+++ governance-20210120/ChangeLog.md
@@ -1,3 +1,13 @@
+2024-11-15 Version: 0.3.4
+- Update API EnrollAccount: add param Tag.
+- Update API EnrollAccount: update param AccountNamePrefix.
+- Update API EnrollAccount: update param DisplayName.
+- Update API GetEnrolledAccount: update response param.
+- Update API ListEvaluationMetricDetails: update response param.
+- Update API ListEvaluationScoreHistory: add param AccountId.
+- Update API RunEvaluation: add param MetricIds.
+
+
 2024-10-12 Version: 0.3.3
 - Update API EnrollAccount: add param Tag.
 - Update API EnrollAccount: update param AccountNamePrefix.